@@ -25,7 +25,7 @@ module Query
 
     def test_it_can_search_extra_attributes
       @history.each do |t|
-        t.update_attributes(tcm_transaction_reference: "x1y2z3",
+        t.update(tcm_transaction_reference: "x1y2z3",
                             generated_filename: "wig77wam")
       end
       # search category, tcm_transaction_reference, generated_filename
